+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # 0.1.0 / FIX ME
 
-## Broken APIs
+## Breaking changes
 
-* The AST node kind :varref has been renamed to :var_ref
+* Node kind :varref has been renamed to :var_ref
+* Node kinds :or, :and, :not have been renamed as :bool_or, :bool_and and :bool_not
 
 # 0.0.1 / 2012-02-16